server/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
dnspython==1.16.0
docstring-parser==0.15
Flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
instructor==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==1.12.0
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==3.11.0
python-dotenv==1.0.1
rich==13.7.0
sniffio==1.3.0
tenacity==8.2.3
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.9.0
Werkzeug==3.0.1
yarl==1.9.4
